requirements.txt
# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep learning
torch>=2.1.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment configuration
python-dotenv>=1.0.0
